Move signature api tests to their own file
